Refactor 헥사고날 아키텍처에 맞게 service 디렉토리를 application으로 명칭 변경해라 - #112
