+++ tictoc-api/src/main/java/tictoc/bid/application/BidQueryService.java
@@ -1,4 +1,4 @@
-package tictoc.bid.service;
+package tictoc.bid.application;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Pageable;
